chore: mask content in root.go config
